+++ services/column-type-getter.js
@@ -57,6 +57,7 @@ function ColumnTypeGetter(databaseConnection, schema, allowWarning = true) {
 
     switch (type) {
       case 'BIT': // NOTICE: MSSQL type.
+      case 'BIT(1)': // NOTICE: MySQL type.
       case 'BOOLEAN':
         return 'BOOLEAN';
       case 'CHARACTER VARYING':
@@ -82,14 +83,14 @@ function ColumnTypeGetter(databaseConnection, schema, allowWarning = true) {
         return 'UUID';
       case 'JSONB':
         return 'JSONB';
-      case 'SMALLINT':
       case 'INTEGER':
       case 'SERIAL':
       case 'BIGSERIAL':
       case (type.match(/^INT.*/i) || {}).input:
+      case (type.match(/^SMALLINT.*/i) || {}).input:
       case (type.match(/^TINYINT.*/i) || {}).input:
         return 'INTEGER';
-      case 'BIGINT':
+      case (type.match(/^BIGINT.*/i) || {}).input:
         return 'BIGINT';
       case (type.match(/FLOAT.*/i) || {}).input:
         return 'FLOAT';
@@ -103,7 +104,7 @@ function ColumnTypeGetter(databaseConnection, schema, allowWarning = true) {
         return 'DOUBLE';
       case 'DATE':
       case 'DATETIME':
-      case 'TIMESTAMP':
+      case (type.match(/^TIMESTAMP.*/i) || {}).input:
       case 'TIMESTAMP WITH TIME ZONE':
       case 'TIMESTAMP WITHOUT TIME ZONE':
         return 'DATE';